requirements.txt
# Core ML/NLP
torch
transformers
peft
sentence-transformers
spacy

# Data + plotting
numpy
pandas
matplotlib
seaborn

# Dimensionality reduction
umap-learn
scikit-learn

# Config
PyYAML

# NLP lexicon
ltk

# Quantization
bitsandbytes
accelerate
